# master/src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Distributed File System master service: node orchestration ve client metadata yönetimi.",
        "title": "DFS Master API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3030",
    "basePath": "/",
    "paths": {
        "/api/acknowledgement": {
            "post": {
                "description": "Node bir chunk'ı başarıyla kaydettiğinde master'a bildirir. Replica sayacı artırılır.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "node"
                ],
                "summary": "Chunk ack al",
                "parameters": [
                    {
                        "description": "Ack bilgileri",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_M1ralai_DFS_src_internal_module_node_dto.AckRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_M1ralai_DFS_src_utils_response.Response-int"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_M1ralai_DFS_src_utils_response.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_M1ralai_DFS_src_utils_response.Response-any"
                        }
                    }
                }
            }
        },
        "/api/file": {
            "delete": {
                "description": "Verilen file_id'ye sahip dosyayı ve tüm chunk'larını siler.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "client"
                ],
                "summary": "Dosyayı sil",
                "parameters": [
                    {
                        "description": "Silme isteği",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_M1ralai_DFS_src_internal_module_client_dto.DeleteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_M1ralai_DFS_src_utils_response.Response-int"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_M1ralai_DFS_src_utils_response.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_M1ralai_DFS_src_utils_response.Response-any"
                        }
                    }
                }
            }
        },
        "/api/file/{id}": {
            "get": {
                "description": "Verilen file_id için dosya bilgisi ve chunk placement'ı döner.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "client"
                ],
                "summary": "Dosya metadata getir",
                "parameters": [
                    {
                        "type": "string",
                        "description": "File UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_M1ralai_DFS_src_utils_response.Response-github_com_M1ralai_DFS_src_internal_module_client_dto_FileResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_M1ralai_DFS_src_utils_response.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_M1ralai_DFS_src_utils_response.Response-any"
                        }
                    }
                }
            }
        },
        "/api/heartbeat": {
            "post": {
                "description": "Node'dan periyodik heartbeat alır, son görülme zamanını ve available space'i günceller.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "node"
                ],
                "summary": "Node heartbeat al",
                "parameters": [
                    {
                        "description": "Heartbeat bilgileri",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_M1ralai_DFS_src_internal_module_node_dto.HeartbeatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_M1ralai_DFS_src_utils_response.Response-int"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_M1ralai_DFS_src_utils_response.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_M1ralai_DFS_src_utils_response.Response-any"
                        }
                    }
                }
            }
        },
        "/api/node": {
            "get": {
                "description": "Master'a kayıtlı tüm node'ları döner.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "node"
                ],
                "summary": "Tüm node'ları listele",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_M1ralai_DFS_src_utils_response.Response-array_github_com_M1ralai_DFS_src_internal_module_node_dto_NodeSaveRequest"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_M1ralai_DFS_src_utils_response.Response-any"
                        }
                    }
                }
            },
            "post": {
                "description": "Bir storage node'unu master'a kaydeder.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "node"
                ],
                "summary": "Yeni node kaydet",
                "parameters": [
                    {
                        "description": "Node kayıt bilgileri",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_M1ralai_DFS_src_internal_module_node_dto.NodeSaveRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_M1ralai_DFS_src_utils_response.Response-string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_M1ralai_DFS_src_utils_response.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_M1ralai_DFS_src_utils_response.Response-any"
                        }
                    }
                }
            }
        },
        "/api/upload": {
            "post": {
                "description": "Bir dosya için chunk placement üretir. Hangi chunk'ın hangi node'lara yazılacağını döner.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "client"
                ],
                "summary": "Dosya yükleme planı oluştur",
                "parameters": [
                    {
                        "description": "Yüklenecek dosya bilgileri",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_M1ralai_DFS_src_internal_module_client_dto.UploadRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_M1ralai_DFS_src_utils_response.Response-github_com_M1ralai_DFS_src_internal_module_client_dto_UploadResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_M1ralai_DFS_src_utils_response.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_M1ralai_DFS_src_utils_response.Response-any"
                        }
                    }
                }
            }
        },
        "/api/user/{user_id}/files": {
            "get": {
                "description": "Verilen user_id'ye ait tüm dosyaları ve chunk placement'larını döner.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "client"
                ],
                "summary": "Kullanıcının dosyalarını listele",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User UUID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_M1ralai_DFS_src_utils_response.Response-github_com_M1ralai_DFS_src_internal_module_client_dto_UserFilesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_M1ralai_DFS_src_utils_response.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_M1ralai_DFS_src_utils_response.Response-any"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_M1ralai_DFS_src_internal_module_client_dto.ChunkLocation": {
            "type": "object",
            "properties": {
                "chunk_id": {
                    "type": "string"
                },
                "nodes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "github_com_M1ralai_DFS_src_internal_module_client_dto.DeleteRequest": {
            "type": "object",
            "required": [
                "file_id",
                "user_id"
            ],
            "properties": {
                "file_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "github_com_M1ralai_DFS_src_internal_module_client_dto.FileResponse": {
            "type": "object",
            "properties": {
                "chunks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_M1ralai_DFS_src_internal_module_client_dto.ChunkLocation"
                    }
                },
                "file_id": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "github_com_M1ralai_DFS_src_internal_module_client_dto.UploadRequest": {
            "type": "object",
            "required": [
                "file_size",
                "filename",
                "user_id"
            ],
            "properties": {
                "file_size": {
                    "type": "integer"
                },
                "filename": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "github_com_M1ralai_DFS_src_internal_module_client_dto.UploadResponse": {
            "type": "object",
            "properties": {
                "chunks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_M1ralai_DFS_src_internal_module_client_dto.ChunkLocation"
                    }
                },
                "file_id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_M1ralai_DFS_src_internal_module_client_dto.UserFilesResponse": {
            "type": "object",
            "properties": {
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_M1ralai_DFS_src_internal_module_client_dto.FileResponse"
                    }
                }
            }
        },
        "github_com_M1ralai_DFS_src_internal_module_node_dto.AckRequest": {
            "type": "object",
            "required": [
                "available_space",
                "chunk_id",
                "id"
            ],
            "properties": {
                "available_space": {
                    "type": "integer"
                },
                "chunk_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "github_com_M1ralai_DFS_src_internal_module_node_dto.HeartbeatRequest": {
            "type": "object",
            "required": [
                "available_space",
                "node_id"
            ],
            "properties": {
                "available_space": {
                    "type": "integer"
                },
                "node_id": {
                    "type": "string"
                }
            }
        },
        "github_com_M1ralai_DFS_src_internal_module_node_dto.NodeSaveRequest": {
            "type": "object",
            "required": [
                "available_space",
                "chunks",
                "node_id"
            ],
            "properties": {
                "available_space": {
                    "type": "integer"
                },
                "chunks": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "node_id": {
                    "type": "string"
                }
            }
        },
        "github_com_M1ralai_DFS_src_utils_response.Response-any": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "github_com_M1ralai_DFS_src_utils_response.Response-array_github_com_M1ralai_DFS_src_internal_module_node_dto_NodeSaveRequest": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_M1ralai_DFS_src_internal_module_node_dto.NodeSaveRequest"
                    }
                },
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "github_com_M1ralai_DFS_src_utils_response.Response-github_com_M1ralai_DFS_src_internal_module_client_dto_FileResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_M1ralai_DFS_src_internal_module_client_dto.FileResponse"
                },
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "github_com_M1ralai_DFS_src_utils_response.Response-github_com_M1ralai_DFS_src_internal_module_client_dto_UploadResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_M1ralai_DFS_src_internal_module_client_dto.UploadResponse"
                },
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "github_com_M1ralai_DFS_src_utils_response.Response-github_com_M1ralai_DFS_src_internal_module_client_dto_UserFilesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_M1ralai_DFS_src_internal_module_client_dto.UserFilesResponse"
                },
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "github_com_M1ralai_DFS_src_utils_response.Response-int": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "github_com_M1ralai_DFS_src_utils_response.Response-string": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}